darwin: detect heterogeneous caches from sysctl perflevels
When we get cpukinds from IOKit, we have to match them
with perflevels (P=0 and E=1 for now) to find out which
cpukinds correspond to sysctl perflevel caches.

If we can't match (may occur if future core types are
different from E and P), we fallback to sysctl perflevels.

Fixes #507

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>