utils: allow loading fsroot/cpuid topologies from a tarball
If -i/--input is given a file ending with .tar.bz2 or .tar.gz
(or --input-format/--if "archive" is given), use archivemount
to fusemount the tarball instead of requiring unpacking it
before load.

We enter the mountpoint, 'umount -l' and 'rmdir' it immediately
so that we don't leak a mountpoint if lstopo crashes.

test-topology.sh under tests/hwloc/{linux,x86} still unpacks
since archivemount is rarely installed.
Specific archivemount tests will be committed separately.

This optional feature is mostly convenient to avoid having
millions of small files in lots of fsroot topologies on disk
in case there's a need to test lstopo manually.
I have 2 millions of them, and my backup tool doesn't like that,
replacing them with tarballs will cleanup all this.

Closes #531

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>