memattrs: heuristics to set NUMA node subtype to DRAM/HBM/SPM/NVM
Internally, we classify NUMA nodes by tier:
1) "UNKNOWN" (usually DRAM)
2) "SPM" (UEFI name "Specific-Purpose Memory") is what Linux exposes
   as "Soft-Reserved" DAX, usually HBM but could be something else.
3) "NVM" (NVDIMMs already detected on Linux through dax/kmem)
4) "GPU" (NVIDIA-only, already detected on Linux, and exposed with subtype "GPUMemory")

If (2) has 2x higher bandwidth than (1), (2) becomes HBM and (1) become DRAM.
If HWLOC_MEMTIERS_GUESS=spm_is_hbm is set in the environment, we don't even
look a the bandwidth.

In the end, we set DRAM/HBM/NVM to NUMA node subtypes.
We keep SPM if we couldn't guess that SPM was HBM.
DRAM isn't set unless there's anything else in the system.

The heuristics is applied at the end of the topology, even when
loading from XML.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>