linux/io: no need for sysfs local_cpus for OSdev locality
Most OSdev have a numa_node attribute in their sysfs hierarchy,
but local_cpus is only available PCI sysfs devices.

OSdev that are related to a PCIdev are attached below that
PCIdev. We don't need to look at numa_node or local_cpus
here. The locality of the PCIdev was obtained earlier by
looking at local_cpus in hwloc_linux_backend_get_pci_busid_cpuset().
numa_node is also available (since 2007) but not used.

OSdev that are NOT related to a PCIdev have no relevant
local_cpus to look at, hence this code was never used.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>