linux/io: rework/fix numa_node attribute in sysfs
For class devices (from /sys/class/foo), the actual device
is pointed by the "device" symlink, there's usually a numa_node
attribute in there.

For bus devices (from /sys/bus/foo/devices), numa_node can be
in the device itself (hmem DAX devices in some Suse 5.3 kernel),
or in its parent (old DAX devices in pre-5.5 kernels),
or in both (newer kernels).

The previous code only supported the parent case. We now
use the device itself first for bus devices (enough in the vast
majority of cases), and then the parent if a flag was given
(for DAX on some old kernels).

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>