linux+lstopo/draw: expose CXL memory device (Type 3) size
Linux support for exposing this memory as NUMA node isn't
totally ready yet (at least for RAM), but we already have
device ram/pmem size in sysfs.

This is a Block device just like DAX for now, we'll work
on a better way to expose all these memory things in 3.0.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>