lstopo: improve the error message in case of unrecognized option
When an option is unrecognized, it gets used as the output filename.
The real filename, if any, then throws an error saying it's an
unrecognized option. Clarify that it couldn't be used as an output
filename because another one was already used.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit 51f3c81e8cf874952047b4fb4553501253e3bc0a)