linux/numa: disable memory attributes and memory-side cache in fake numa
HMAT information isn't updated when fake numa is enabled,
so ignore it for now.

In the uniform splitting fake numa case, we could update HMAT info
since node #K is replaced with nodes #KN-#(K+1)N-1.
In theory, the Linux kernel should fix HMAT info to match the
fakenuma config, but it looks like it's not going to happen soon,
we'll see later if we need/want to fix it in hwloc.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>