tests/backends.c: fix a failure when /sys isn't available on Linux on non-x86
Since 2.9, Linux discovery aborts if /sys isn't available.
Usually it will fallback to the x86 backend... except on non-x86 platforms obviously
where it would fallback to "noos" instead.

The hwloc_backends.c test verified that the "Backend" info attribute was the same
in the XML-loaded topology and the original topology, but it did not handle the
case where that info attribute did not exist at all, which is the case when
only "noos" is used.

This is first reported by Simon South on https://issues.guix.gnu.org/61493
Thanks to Ludovic Courtes for forwarding to me.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit 0780a3b9002c3567f3a8041a62f0d81bbddfd027)