tests/{linux,xml,x86,x86+linux,linux/allowed}: allow to run tests in an optional extra subdirectory
If extra/extra-tests.sh exists, call it with our TEST_LOG_COMPILER as parameter.

Useful for additional testing that we can't or don't want to put in the repo
(confidential, huge files, ...).
Mostly useful for x86 and linux but supporting others is easy.

Contrary to #191, this does not touch automake stuff at all,
everything is managed by the outside subdirectory script
(which basically iterates over tests and pass them to out TEST_LOG_COMPILER).

Closes #191

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>