restrict: fix when doing nothing except because of REMOVE_CPULESS
If we restrict to the whole topology cpuset, we usually do nothing
... except if REMOVE_CPULESS is passed: some NUMA nodes with
empty cpusets will get removed.
We had a buggy assertion because of this, raised in #575.

Moreover, fix the way we mark the object has "modified". With
the current code, the parent wouldn't be considered modified,
hence we wouldn't recurse in children to actually remove those
NUMA nodes. And that caused the nodeset/complete_nodeset to
remain untouched while allowed_nodeset was reduced in the caller.

Also fix the corresponding case when restricting by nodeset
and removing memory-less CPUs.

Thanks to Mark Grondona for the report.

Closes #574

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit 22a193101bbeff92f6371d71eca6ad87d3b1a7d9)