x86: fix/cleanup the breaking-out of the topology enumeration loops
Intel and AMD specify different conditions for breaking out of the
loop on CPUID 0xb/0x1f and 0xb/0x80000026 leaves respectively.
And Zhaoxin behaves slightly differently from Intel.
We cannot factorize the break condition as we did in commit
20c794500c810686b0c0099b1e099c7461ff9f6b, just respect the actual specs
depending on whether we are on AMD or not.

This fixes Dies on Zhaoxin KH-4000 CPUs (that 0x1f level was ignored
because it reports 0 in the number of threads in next level).
These dies aren't visible in any release yet since 0x1f wasn't enabled
on Zhaoxin (enabled in commit 4f89de1f902d27ebf3844632d5a7a4412ba5e450).

Refs #593

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>