levelzero: disable ZES_ENABLE_SYSMAN=1 when zesInit() is available
When zesInit() is available (starting in specs 1.5), we don't need
all the stuff to set ZES_ENABLE_SYSMAN=1 in the environment anymore.

This assumes that zesInit() works for real, which was not the case
in first compute-runtime releases (it returned "unsupported").
Those releases won't be supported anymore because they'd need
a runtime-check at configure time to find out that the ZES_ENABLE_SYSMAN=1
stuff must be enabled.

Only use a compile-check to switch between zesInit() and
ZES_ENABLE_SYSMAN=1, and tell users to avoid those L0 runtime
versions where zesInit() fails.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>