levelzero: use zesInit() when available
When zesInit() is available (starting in specs 1.5), call it
to further reduce problems with ZES_ENABLE_SYSMAN=1 not
being set in the environment.

The current implementation of zesInit() in the oneAPI compute runtime
returns an "unsupported feature" 0x78000003 error
(or return success and do nothing if ZES_ENABLE_SYSMAN=1)
but hopefully things will improve quickly.

Once zesInit() will be widely available/implemented, we'll
consider disabling the ZES_ENABLE_SYSMAN=1 stuff if zesInit()
is found at configure time, and later remove it.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>