darwin: fix IOKit warnings about const char * being passed in arrays
GCC11 emits warnings [-Wstringop-overread] when passing short C strings in
io_name/string_t in the IOKit API because those types are 128/512 char arrays.
It's not clear if these warnings are really valid but they are not too hard to fix
(we cannot cast into an array type, so just declare a variable of that array type earlier).

By the way, update the actual definition of io_name/string_t in our port test headers.

Thanks to Ralph Castain for the report.

Fixes #599

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>