allow: now allowed on shmem-adopted topologies
A shmem topology may now be shared between different cgroups, one only has
to call hwloc_topology_allow(LOCAL) to update the list of allowed CPUs and memory
after adopting the shmem topo.
The source process must have built the original topology with INCLUDE_DISALLOWED.

This was disabled previously because allow(LOCAL) is supposed to add "LinuxCgroup"
to the info attrs, which was not possible until we added topo infos in 3.0.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>