utils/calc: fix searching CPU objects below heterogeneous NUMA
socket:0.numa:0.core:0 looks for first core whose cpuset and nodeset
are included in socket:0.numa:0. This fails on heterogeneous memory
machines since this core may have 2 local NUMAs in its nodeset,
hence not included in socket:0.numa:0's nodeset.

Fix this by checking if the nodeset intersects instead of being included.

Note that the empty bitmap is isincluded in any bitmap, while it doesn't
intersect with anything, hence we have to check empty nodesets too.

Thanks to Antoine Morvan for the report.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>