test-gather-topology.sh: Disable test when cross-compiling
In case one uses binfmt to be able to run the cross-built library &
programs, the content of /proc/cpuinfo is supposed to be the guest cpu
content, while cat will have gotten the host cpu content.

Actually at the moment qemu erroneously gives the host cpu content, see

https://gitlab.com/qemu-project/qemu/-/issues/2004

but once that will get fixed, cross-compiled gather-topology testing will
fail (as expected), so disable it in that case.

Refs #638