API: add hwloc_obj_set_subtype()
hwloc_topology_insert_group() says explicitly that group->subtype may
be defined by the user. However it will be freed by hwloc's internal
call to free(), which may be incompatible to how the caller allocated
that subtype.
Now recommend the use of hwloc_obj_set_subtype() instead, it will take
care of duplicating the user-given given subtype.

Thanks to HadrienG2 for the report.

Closes #642.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>