utils/calc+info: rework --best-memattr to allow multiple nodes
--best-memattr was very strict when selecting best nodes.
The node had to be the best for the entire input CPUs.
On a dual-socket machine with HBM in each socket, each HBM
is the best local but not the best remote. Hence we'd report
no best for the entire machine.

Now we return both HBM for the entire machine by default,
but may go back to the previous behavior by adding ",strict"
after the attribute name.

Adding ",default" also allows to return all nodes if no best
was found (if no attribute values are found).

Thanks to Antoine Morvan for the report.

Refs #652

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit 313de564cd1ee2a2f3a1fa208316f523b3da0ca2)

(with compile fixes forgotten in intermediate commits in master)