windows: fix dll filenames when UNICODE is enabled
LoadLibrary() and GetModuleHandle() failed when UNICODE
is enabled because the filename was not properly encoded.
Use the TEXT() macro to fix the encoding.

Most dynamic function pointers would be NULL, and we'd
fail to discover many features and to bind.

Thanks to Martin for the report.

Closes #655

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit b4ce80732bb9abd699f68229db3c489669777351)