core+linux: generalize and improve Die filtering vs Packages
The Linux backend ignored Dies that are identical to packages because
the kernel may expose that even if the hardware does not support Dies.
Windows might do the same (see #480 but no reply from Microsoft yet).

Move the filtering to the core and only apply it if all Dies are
identical to their Package.

Implemented by forcing KEEP_STRUCTURE filtering only between Dies
and Packages.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>