utils/calc: rework --cpuset-output-format for the systemd-dbus-api
The new --nodeset-output-format does --cpuset-output-format and --nodeset-output.
It's not listed in usage to avoid using too many lines for cpuset formats,
but it's listed in the manpage.

The systemd-dbus-api doesn't hardwire the AllowedCPUs prefix anymore
so that it is also available for nodeset outputs, the final user will
add any of
- AllowedCPUs
- StartupAllowedCPUs
- AllowedMemoryNodes
- StartupAllowedMemoryNodes
before the output of the `hwloc-calc` command.

This breaks the behavior introduced in 2.11.2 but the (very likely)
only user is the author of this commit.

Signed-off-by: Pierre Neyron <pierre.neyron@imag.fr>
Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit ee03ef259f1d03e7d9df932d9cf3ab3757532416)