levelzero: separate ZE and ZES device handles
They are the same for now because we use the legacy way
to enable Sysman, which allows to cast between handles,
but it won't be the case anymore soon.

By the way, clear an array on error to help coverity.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>
(cherry picked from commit e0a195b1e90343b443f01d041e1bf243777aa212)