rsmi: fix the PCI locality of partitioned devices
rsmi_dev_pci_id_get() returns the GPU "partition ID" inside the PCI BDF function,
but this virtual function isn't actually exposed to the OS.
See ROCm/rocm_smi_lib#208 for details.

When hwloc fails to find the corresponding PCI device (usually gets the above bridge instead),
if the BDF function is > 0, get the RSMI partition ID, compare it with the BDF function,
and try to get the PCI device with func = 0 instead.

rsmi_dev_partition_id_get() was only added in ROCm 6.2, so configure-check it.

Thanks to Edgar Leon for reporting and debugging the issue.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>