doxygen: ignore the static keyword
Our (many) static inline helpers are shown with the static prefix in the list
of fonctions of modules, and later with the static prefix and static+inline
labels in their own subsection. See doxygen/doxygen#11419

Users are confused by this static without inline prefix, and we don't have
any use of static outside static inline in the doc, so just ignore static
entirely.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>