misc: reactivate GNUC conditional
This avoids use of the custom hwloc_flsl_manual() implementation if a
recent GCC is in use.
It is not clear why the erroneous underscores, with which the macro does
not seem to be defined anywhere, were introduced, as the commit introducing
the line only references an SVN import. Older source releases however do
not contain the change.

This was discovered as hwloc_flsl_manual() returns zero and as a result a
division by zero exception with recent versions of GCC on certain CPU
configurations, whereas the native functionality behaves correctly on
both tested versions.
The manual implementation being defective is technically still an issue,
however the logic under the __GNUC__ conditional is deemed to be preferable
in any case.

Fixes: 3ac17a230d2a ("topo_cpuset_last")
Link: https://bugzilla.opensuse.org/show_bug.cgi?id=1236038
Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>