test-gather-topology: disable Misc objects
MemoryModules are only reported when DMI sysfs is enabled
and accessible, ie as root.
But we don't gather DMI info in the test, hence a run as root
fails because the default lstopo gets them while lstopo on the
dump doesn't.
Just ignore Misc objects since this doesn't matter much here.

Thanks to Andrew Phelps for the report.

Closes #719

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>