Do not treat INTERSECT_LOCALITY as a superset of LARGER and SMALLER
To be fair, it is a superset for almost all cpusets. But there is one exception to this rule, which is the empty cpuset.

- It passes the LARGER_LOCALITY test for all NUMA nodes.
- It passes the INTERSECT_LOCALITY for no NUMA node.

As a result, we get to a situation where for this particular cpuset, adding an `hwloc_local_numanode_flag_e` (INTERSECT_LOCALITY) results in a smaller number of NUMA nodes being matched, which violates normal user expectations about `hwloc_get_local_numanode_objs()`.

(cherry picked from commit 0874793771c939a49c73241bdd1edf379ef6efb1)