doc/readthedocs: move the doxygen doc to a subdirectory and redirect to it
Sphinx overwrites the doxygen index.
The proper solution (and possibly nicer) is to use things like breathe,
exhale or doxysphinx to actually integrate doxygen in sphinx,
but that's a lot of work.

Hence rather move the doxygen generated HTML doc to a subdirectory
and make a dummy Sphinx index with only a link to the doxygen index.

And we enable a (forced) redirection in the readthedocs config (on the website)
so that the sphinx index isn't even visible.

Somehow, readthedocs manages to put its own stuff on top of our
doxygen pages, including the bottom-left menu for switching between versions.
The only issue I found is that the warning about
"latest version not being stable" is hidden behind the doxygen title bar.

By the way, we don't use autotools at all on readthedocs,
hence we have to duplicate the fig2dev command to generate diagram.png.
Might be better to just prebuild diagram.{png,eps} as we do for lstopo
outputs since we pretty much never modify diagram.fig.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>