Add SPDX license headers and fix some copyright headers
* Add "SPDX-License-Identifier: BSD-3-Clause" in all files where it matters
* Uniformize headers with SPDX first, then copyright lines,
  then "See COPYING in top-level directory." (which is used by
  contrib/update-my-copyright.pl to find where to add copyrights)
* Remove $COPYRIGHT$ and $HEADER$, those were used by contrib/update-my-copyright.pl
  in a few files, not needed anymore.
* Update COPYING with notes about a few imported files with different licenses

For the record, our license isn't exactly BSD-3-Clause because we say
"the authors" instead of "the copyright holders and contributors".
According to lawyers it doesn't matter much, but cannot be fixed
without asking all copyright holders (which would be very hard).
It's not clear where the small difference comes from.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>