windows: fix the plugin loading code with cygwin
It builds but I couldn't get it to load yet.
When I put a CMake-built plugin in the right path, it finds it but
loading fails with "the specific module could not be found".

When actually enabling this in non-CMake builds, we'll need
and to define HWLOC_PLUGINS_PATH in the MSVC config.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>