API: remove "page_type" from NUMA node attributes
On !Linux, we may have a huge page size but not the number
of pages for each size.

On Linux, the number of hugepages is available by NUMA node but
it does not make much sense since the kernel has been able to create
huge pages transparently for a while. It only made sense in the past
when hugepages had to be force-created manually by the admin.

We'll just export PageSizes as a new topo info.

This breaks the API/ABI but it's already broken in 3.0 anyway.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>