pci: merge the forced and normal locality lists
Forced localities are immediately inserted in the common list,
and they are extended to neighbor buses in same PCI hierarchies.

The list is sorted by domain/bus.
To avoid traversing the list (usually small anyway), we
remember the last used one since buses are inserted in order.

Add some debug messages to list forced and final localities,
and debug checks to verify the list order.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>